add seo component
adds the seo component to the app. it's responsible for setting up the metadata for the app, such as the title, description, keywords, etc., as well as the og tags and twitter tags. it also adds a link to the favicon and apple touch icon. finally, it also adds a google analytics tag to the page.

the following changes were made:

- added seo component to `/src/components/seo/index.js`
- added `seo` component to `/src/pages/_document.js`
- added google analytics tag to `/src/pages/_document.js`