update seo title
updates the title of the seo component to use the summarizer instead of italy news summarizer. this change will ensure that the title of the page is consistent with the rest of the website.

```diff
diff --git a/src/components/seo/index.js b/src/components/seo/index.js
index 2fdfbb534fb3..c50a42032fd1 100644
--- a/src/components/seo/index.js
+++ b/src/components/seo/index.js
@@ -2,7 +2,7 @@ import head from \"next/head\";

 const seo = ({ description, keywords, title }) => (
   <>
-    <title>{title} | italy news summarizer</title>
+    <title>{title} | summarizer</title>
     <meta name=\"description\" content={description} />
     <meta name=\"keywords\" content={keywords?.join(\", \")} />
     <meta property=\"og:type\" content=\"website\" />
```