add promisehandler function in utils/index.js
adds a new function called `promisehandler` to the `utils/index.js` file. this function takes a promise and returns either the data on success or an error on failure. the function returns an array with the first element being either an error or `null` and the second element being either the data or the error.

adds the following code to `utils/index.js`:

```js
/**
 *
 * takes a promise and returns the promise's data on success or
 * an error on failure.
 *
 * @param {promise} promise - the promise to be handled
 *
 * @returns {array} - an array with the first element being either
 * an error or null and the second element being either the data
 * or the error.
 */
const promisehandler = promise => {
  return promise.then(data => [null, data]).catch(err => [err]);
};
```